km tools v1.1
Bare bones of the plugin, added toolbar and menu functionality, cleaned
up the code a bit. I'm still new to ruby and I have to figure out how
this plugin will actually interact with the UI and the model. The goal
is to create a tool that will allow the user to scale a selected object
to specified absolute dimensions instead of scaling based on a relative
percentage.